refactor: icon 컴포넌트 속성 단일화 및 스타일 적용 방식 수정 (#7)
- width, height => size로 변경
- iconClassName 카멜케이스 방식으로 이름 재수정
- menuData.ts 파일을 constants 폴더 내로 이동
