+++ src/components/Icon.tsx
@@ -3,23 +3,18 @@ import { cn } from '@utils/cn'
 
 export default function Icon({
   icon: ImgIcon,
-  width = 18,
-  height = 18,
+  size = 18,
   color = 'text-primary-600',
   className,
-  IconClassName,
+  iconClassName,
 }: IconProps) {
-  const formatSizeValue = (value: number | string) =>
-    typeof value === 'number' ? `${value}px` : value
-
   return (
-    <div className={cn(`flex items-center justify-center`, className)}>
+    <div className={cn(`flex items-center justify-center`, className)} style={{
+          width: `${size}px`,
+          height:`${size}px`,
+        }}>
       <ImgIcon
-        className={cn(color, IconClassName)}
-        style={{
-          width: formatSizeValue(width),
-          height: formatSizeValue(height),
-        }}
+        className={cn(color, iconClassName)}
       />
     </div>
   )
